feat: improve update recording logic in backend
The idnode of a tvheadend recording is now updated via a patch.
The existing recording will be fetched from tvheadend and merged
with the update options from the request.

This provides a safer way to update a recording, because the previous
way could have destroyed the entire recording when required data were
not provided.